fs_mgr: Honor fstab partition length formatting to ext4
* Change fstab_rec length type to uint64_t
* Parse lenght with parse_size ( allows to use k/m/g suffixes )

+++ fs_mgr/include_fstab/fstab/fstab.h
@@ -46,7 +46,7 @@ struct fstab_rec {
     char* key_loc;
     char* key_dir;
     char* verity_loc;
-    long long length;
+    uint64_t length;
     char* label;
     int partnum;
     int swap_prio;
